+++ remove_arch.ps1
@@ -1,1 +1,0 @@
-Remove-Item "d:\Foundry\Data\systems\cogwheel-syndicate\packs\archetypes" -Recurse -Force